- Se solucionan los problemas de escritura del HDF5 cuando la configuracion es muy grande - Se sustituye el tipo de dato para las dimensiones del reticulo de uint a size_t, para admitir grandes configuraciones

+++ include/fftCode.h
@@ -1,3 +1,3 @@
-void	initFFT	(void *m, void *m2, const uint n1, const uint Tz, FieldPrecision prec, bool lowmem);
+void	initFFT	(void *m, void *m2, const size_t n1, const size_t Tz, FieldPrecision prec, bool lowmem);
 void	runFFT	(int sign);
 void	closeFFT();
