+++ lib/propXeon.cpp
@@ -1,3 +1,4 @@
+#include<cstdio>
 #include<cmath>
 #include"scalarField.h"
 #include"enum-field.h"
@@ -38,7 +39,7 @@
 #include<cstdio>
 #ifndef	__MIC__
 #ifdef	__AVX__
-void printFloat(uint idx, uint con, __m256 dat)
+void printFloat(size_t idx, size_t con, __m256 dat)
 {
 	if (idx == con) {
 		float caca[8];
@@ -47,7 +48,7 @@ void printFloat(uint idx, uint con, __m256 dat)
 	}
 }
 
-void printDouble(uint idx, uint con, __m256d dat)
+void printDouble(size_t idx, size_t con, __m256d dat)
 {
 	if (idx == con) {
 		double caca[4];
@@ -58,7 +59,7 @@ void printDouble(uint idx, uint con, __m256d dat)
 
 #else
 
-void printFloat(uint idx, uint con, __m128 dat)
+void printFloat(size_t idx, size_t con, __m128 dat)
 {
 	if (idx == con) {
 		float caca[4];
@@ -67,7 +68,7 @@ void printFloat(uint idx, uint con, __m128 dat)
 	}
 }
 
-void printDouble(uint idx, uint con, __m128d dat)
+void printDouble(size_t idx, size_t con, __m128d dat)
 {
 	if (idx == con) {
 		double caca[2];
@@ -78,7 +79,7 @@ void printDouble(uint idx, uint con, __m128d dat)
 #endif
 #else
 __attribute__((target(mic)))
-void printFloat(uint idx, uint con, __m512 dat)
+void printFloat(size_t idx, size_t con, __m512 dat)
 {
 	if (idx == con) {
 		static float __attribute((aligned(64))) caca[16];
@@ -89,7 +90,7 @@ void printFloat(uint idx, uint con, __m512 dat)
 }
 
 __attribute__((target(mic)))
-void printDouble(uint idx, uint con, __m512d dat)
+void printDouble(size_t idx, size_t con, __m512d dat)
 {
 	if (idx == con) {
 		static double  __attribute((aligned(64))) caca[8];
@@ -102,9 +103,9 @@ void printDouble(uint idx, uint con, __m512d dat)
 __attribute__((target(mic)))
 #endif
 void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, void * __restrict__ m2_, double *z, const double dz, const double c, const double d,
-			    const double ood2, const double LL, const double nQcd, const uint Lx, const uint Vo, const uint Vf, FieldPrecision precision)
+			    const double ood2, const double LL, const double nQcd, const size_t Lx, const size_t Vo, const size_t Vf, FieldPrecision precision)
 {
-	const uint Sf = Lx*Lx;
+	const size_t Sf = Lx*Lx;
 
 	if (precision == FIELD_DOUBLE)
 	{
@@ -140,8 +141,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const double zQ = 9.*pow(zR, nQcd+3.);
 
 #ifdef	__MIC__
-		const uint XC = (Lx<<2);
-		const uint YC = (Lx>>2);
+		const size_t XC = (Lx<<2);
+		const size_t YC = (Lx>>2);
 
 		const double __attribute__((aligned(Align))) z2Aux[8]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[8]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0. };	// Only real part
@@ -151,8 +152,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const double __attribute__((aligned(Align))) dzcAux[8] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 		const double __attribute__((aligned(Align))) dzdAux[8] = { dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd };
 #elif	defined(__AVX__)
-		const uint XC = (Lx<<1);
-		const uint YC = (Lx>>1);
+		const size_t XC = (Lx<<1);
+		const size_t YC = (Lx>>1);
 
 		const double __attribute__((aligned(Align))) z2Aux[4]  = {-z2,-z2,-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[4]  = { zQ, 0., zQ, 0. };	// Only real part
@@ -162,8 +163,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const double __attribute__((aligned(Align))) dzcAux[4] = { dzc, dzc, dzc, dzc };
 		const double __attribute__((aligned(Align))) dzdAux[4] = { dzd, dzd, dzd, dzd };
 #else
-		const uint XC = Lx;
-		const uint YC = Lx;
+		const size_t XC = Lx;
+		const size_t YC = Lx;
 
 		const double __attribute__((aligned(Align))) z2Aux[2]  = {-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[2]  = { zQ, 0. };	// Only real part
@@ -187,12 +188,12 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 			_MData_ tmp, mel, mPx, mPy;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
-				uint X[2], idxPx, idxMx, idxPy, idxMy, idxPz, idxMz, idxP0;
+				size_t X[2], idxPx, idxMx, idxPy, idxMy, idxPz, idxMz, idxP0;
 
 				{
-					uint tmi = idx/XC, tpi;
+					size_t tmi = idx/XC, tpi;
 
 					tpi = tmi/YC;
 					X[1] = tmi - tpi*YC;
@@ -330,8 +331,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const float zQ = 9.*powf(zR, nQcd+3.);
 
 #ifdef	__MIC__
-		const uint XC = (Lx<<3);
-		const uint YC = (Lx>>3);
+		const size_t XC = (Lx<<3);
+		const size_t YC = (Lx>>3);
 
 		const float __attribute__((aligned(Align))) z2Aux[16]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2};
 		const float __attribute__((aligned(Align))) zQAux[16]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0.};
@@ -341,8 +342,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const float __attribute__((aligned(Align))) dzcAux[16] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 		const float __attribute__((aligned(Align))) dzdAux[16] = { dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd };
 #elif	defined(__AVX__)
-		const uint XC = (Lx<<2);
-		const uint YC = (Lx>>2);
+		const size_t XC = (Lx<<2);
+		const size_t YC = (Lx>>2);
 
 		const float __attribute__((aligned(Align))) z2Aux[8]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2 };
 		const float __attribute__((aligned(Align))) zQAux[8]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0. };
@@ -352,8 +353,8 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 		const float __attribute__((aligned(Align))) dzcAux[8] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 		const float __attribute__((aligned(Align))) dzdAux[8] = { dzd, dzd, dzd, dzd, dzd, dzd, dzd, dzd };
 #else
-		const uint XC = (Lx<<1);
-		const uint YC = (Lx>>1);
+		const size_t XC = (Lx<<1);
+		const size_t YC = (Lx>>1);
 
 		const float __attribute__((aligned(Align))) z2Aux[4]  = {-z2,-z2,-z2,-z2 };
 		const float __attribute__((aligned(Align))) zQAux[4]  = { zQ, 0., zQ, 0. };
@@ -377,12 +378,12 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 			_MData_ tmp, mel, mPx, mPy, mMx;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
-				uint X[2], idxMx, idxPx, idxMy, idxPy, idxMz, idxPz, idxP0;
+				size_t X[2], idxMx, idxPx, idxMy, idxPy, idxMz, idxPz, idxP0;
 
 				{
-					uint tmi = idx/XC, itp;
+					size_t tmi = idx/XC, itp;
 
 					itp = tmi/YC;
 					X[1] = tmi - itp*YC;
@@ -499,11 +500,11 @@ void	propagateKernelXeon(const void * __restrict__ m_, void * __restrict__ v_, v
 	}
 }
 
-void	propagateXeon	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const uint Lx, const uint V, const uint S, FieldPrecision precision)
+void	propagateXeon	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision)
 {
 #ifdef USE_XEON
 	const int  micIdx = commAcc(); 
-	const uint ext = V + S;
+	const size_t ext = V + S;
 	const double ood2 = 1./delta2;
 	double *z = axionField->zV();
 
@@ -567,7 +568,7 @@ void	propagateXeon	(Scalar *axionField, const double dz, const double delta2, co
 #endif
 }
 
-void	propagateCpu	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const uint Lx, const uint V, const uint S, FieldPrecision precision)
+void	propagateCpu	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision)
 {
 	const double ood2 = 1./delta2;
 	double *z = axionField->zV();
@@ -604,7 +605,7 @@ void	propagateCpu	(Scalar *axionField, const double dz, const double delta2, con
 #ifdef USE_XEON
 __attribute__((target(mic)))
 #endif
-void	updateMXeon(void * __restrict__ m_, const void * __restrict__ v_, const double dz, const double d, const uint Vo, const uint Vf, const uint Sf, FieldPrecision precision)
+void	updateMXeon(void * __restrict__ m_, const void * __restrict__ v_, const double dz, const double d, const size_t Vo, const size_t Vf, const size_t Sf, FieldPrecision precision)
 {
 	if (precision == FIELD_DOUBLE)
 	{
@@ -644,7 +645,7 @@ void	updateMXeon(void * __restrict__ m_, const void * __restrict__ v_, const dou
 			register _MData_ mIn, vIn, tmp;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
 #ifdef	__MIC__
 				vIn = opCode(load_pd, &v[2*(idx-Sf)]);
@@ -699,10 +700,10 @@ void	updateMXeon(void * __restrict__ m_, const void * __restrict__ v_, const dou
 		#pragma omp parallel default(shared)
 		{
 			register _MData_ mIn, vIn, tmp;
-			register uint idxP0, idxMz;
+			register size_t idxP0, idxMz;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
 				idxP0 = idx << 1;
 				idxMz = (idx - Sf) << 1;
@@ -728,7 +729,7 @@ void	updateMXeon(void * __restrict__ m_, const void * __restrict__ v_, const dou
 __attribute__((target(mic)))
 #endif
 void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z, const double dz, const double c, const double ood2,
-		    const double LL, const double nQcd, const uint Lx, const uint Vo, const uint Vf, const uint Sf, FieldPrecision precision)
+		    const double LL, const double nQcd, const size_t Lx, const size_t Vo, const size_t Vf, const size_t Sf, FieldPrecision precision)
 {
 	if (precision == FIELD_DOUBLE)
 	{
@@ -758,8 +759,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const double zQ = 9.*pow(zR, nQcd+3.);
 		const double dzc = dz*c;
 #ifdef	__MIC__
-		const uint XC = (Lx<<2);
-		const uint YC = (Lx>>2);
+		const size_t XC = (Lx<<2);
+		const size_t YC = (Lx>>2);
 
 		const double __attribute__((aligned(Align))) z2Aux[8]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[8]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0. };	// Only real part
@@ -768,8 +769,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const double __attribute__((aligned(Align))) d2Aux[8]  = { ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2 };
 		const double __attribute__((aligned(Align))) dzcAux[8] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 #elif	defined(__AVX__)
-		const uint XC = (Lx<<1);
-		const uint YC = (Lx>>1);
+		const size_t XC = (Lx<<1);
+		const size_t YC = (Lx>>1);
 
 		const double __attribute__((aligned(Align))) z2Aux[4]  = {-z2,-z2,-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[4]  = { zQ, 0., zQ, 0. };	// Only real part
@@ -778,8 +779,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const double __attribute__((aligned(Align))) d2Aux[4]  = { ood2, ood2, ood2, ood2 };
 		const double __attribute__((aligned(Align))) dzcAux[4] = { dzc, dzc, dzc, dzc };
 #else
-		const uint XC = Lx;
-		const uint YC = Lx;
+		const size_t XC = Lx;
+		const size_t YC = Lx;
 
 		const double __attribute__((aligned(Align))) z2Aux[2]  = {-z2,-z2 };
 		const double __attribute__((aligned(Align))) zQAux[2]  = { zQ, 0. };	// Only real part
@@ -800,12 +801,12 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 			_MData_ tmp, mel, mPx, mPy;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
-				uint X[2], idxPx, idxMx, idxPy, idxMy, idxPz, idxMz, idxP0;
+				size_t X[2], idxPx, idxMx, idxPy, idxMy, idxPz, idxMz, idxP0;
 
 				{
-					uint tmi = idx/XC, tpi;
+					size_t tmi = idx/XC, tpi;
 
 					tpi = tmi/YC;
 					X[1] = tmi - tpi*YC;
@@ -934,8 +935,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const float zQ = 9.*powf(zR, nQcd+3.);
 		const float dzc = dz*c;
 #ifdef	__MIC__
-		const uint XC = (Lx<<3);
-		const uint YC = (Lx>>3);
+		const size_t XC = (Lx<<3);
+		const size_t YC = (Lx>>3);
 
 		const float __attribute__((aligned(Align))) z2Aux[16]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2};
 		const float __attribute__((aligned(Align))) zQAux[16]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0., zQ, 0.};
@@ -944,8 +945,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const float __attribute__((aligned(Align))) d2Aux[16]  = { ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2 };
 		const float __attribute__((aligned(Align))) dzcAux[16] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 #elif	defined(__AVX__)
-		const uint XC = (Lx<<2);
-		const uint YC = (Lx>>2);
+		const size_t XC = (Lx<<2);
+		const size_t YC = (Lx>>2);
 
 		const float __attribute__((aligned(Align))) z2Aux[8]  = {-z2,-z2,-z2,-z2,-z2,-z2,-z2,-z2 };
 		const float __attribute__((aligned(Align))) zQAux[8]  = { zQ, 0., zQ, 0., zQ, 0., zQ, 0. };
@@ -954,8 +955,8 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 		const float __attribute__((aligned(Align))) d2Aux[8]  = { ood2, ood2, ood2, ood2, ood2, ood2, ood2, ood2 };
 		const float __attribute__((aligned(Align))) dzcAux[8] = { dzc, dzc, dzc, dzc, dzc, dzc, dzc, dzc };
 #else
-		const uint XC = (Lx<<1);
-		const uint YC = (Lx>>1);
+		const size_t XC = (Lx<<1);
+		const size_t YC = (Lx>>1);
 
 		const float __attribute__((aligned(Align))) z2Aux[4]  = {-z2,-z2,-z2,-z2 };
 		const float __attribute__((aligned(Align))) zQAux[4]  = { zQ, 0., zQ, 0. };
@@ -976,12 +977,12 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 			_MData_ tmp, mel, mPx, mPy, mMx;
 
 			#pragma omp for schedule(static)
-			for (uint idx = Vo; idx < Vf; idx += step)
+			for (size_t idx = Vo; idx < Vf; idx += step)
 			{
-				uint X[2], idxMx, idxPx, idxMy, idxPy, idxMz, idxPz, idxP0;
+				size_t X[2], idxMx, idxPx, idxMy, idxPy, idxMz, idxPz, idxP0;
 
 				{
-					uint tmi = idx/XC, itp;
+					size_t tmi = idx/XC, itp;
 
 					itp = tmi/YC;
 					X[1] = tmi - itp*YC;
@@ -1095,11 +1096,11 @@ void	updateVXeon(const void * __restrict__ m_, void * __restrict__ v_, double *z
 	}
 }
 
-void	propLowMemXeon	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const uint Lx, const uint V, const uint S, FieldPrecision precision)
+void	propLowMemXeon	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision)
 {
 #ifdef USE_XEON
 	const int  micIdx = commAcc(); 
-	const uint ext = V + S;
+	const size_t ext = V + S;
 	const double ood2 = 1./delta2;
 	double *z = (double *) __builtin_assume_aligned((void *) axionField->zV(), Align);
 
@@ -1188,7 +1189,7 @@ void	propLowMemXeon	(Scalar *axionField, const double dz, const double delta2, c
 
 #endif
 }
-void	propLowMemCpu	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const uint Lx, const uint V, const uint S, FieldPrecision precision)
+void	propLowMemCpu	(Scalar *axionField, const double dz, const double delta2, const double LL, const double nQcd, const size_t Lx, const size_t V, const size_t S, FieldPrecision precision)
 {
 	const double ood2 = 1./delta2; 
 	double *z = axionField->zV();
